+++ next.config.js
@@ -4,11 +4,6 @@ const path = require('path')
 
 module.exports = {
   reactStrictMode: true,
-  swcMinify: true,
-  experimental: {
-    styledComponents: true
-  },
-  output: { path: path.resolve(__dirname, 'static') },
   webpack: (config, options) => {
     config.module.rules.push({
       test: /\.pdf/,
