+++ src/components/generator/form/Form.tsx
@@ -8,8 +8,22 @@ import { formAtom } from '../../../atoms/form'
 import { resumeAtom } from '../../../atoms/resume'
 import { colors, sizes } from '../../../theme'
 import { FormValues } from '../../../types/form'
-import { generateResume } from '../../../api/generateResume'
 import { progressAtom } from '../../../atoms/progress'
+
+async function generateResume(formData: FormValues): Promise<string> {
+  const pdfResponse = await fetch('/api/generate/pdf', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify(formData)
+  })
+
+  const pdfBlob = await pdfResponse.blob()
+  const pdfUrl = URL.createObjectURL(pdfBlob)
+
+  return pdfUrl
+}
 
 const StyledForm = styled.form`
   display: flex;
