--- src/server.js
@@ -1,29 +1,0 @@
-const next = require('next')
-const Koa = require('koa')
-const Router = require('@koa/router')
-
-const port = Number(process.env.PORT) || 4000
-const dev = process.env.NODE_ENV !== 'production'
-const app = next({ dev })
-const handle = app.getRequestHandler()
-
-app.prepare().then(() => {
-  const server = new Koa()
-  const router = new Router()
-
-  router.all('(.*)', async (ctx) => {
-    await handle(ctx.req, ctx.res)
-    ctx.respond = false
-  })
-
-  server.use(async (ctx, next) => {
-    ctx.res.statusCode = 200
-    await next()
-  })
-
-  server.use(router.routes())
-
-  server.listen(port, () => {
-    console.log(`> Ready on http://localhost:${port}`)
-  })
-})
