--- typings/env.d.ts
@@ -1,1 +1,0 @@
-declare module '*.pdf'
